Adds Home Page and Pug Templating
This commit adds an index.pug in the views directory which
describes how the service works and links to github. Additionally,
routes were updated to serve this page, bootstrap was installed,
a static express route to /scripts was created to link the dist
files for bootstrap. Domain and protocol variables were passed to
the pug index so the page is relative to wherever it's being
served.